requirements.txt
pyfrotz
networkx
fastapi
uvicorn[standard]
websockets
openai
anthropic
google-genai
pytest
